First Commit
The initial version of Parrondo's Paradox, written in C#. Note: in
order to aptly test the paradox, the coinToss variable must be set to
rand.Next(1, 1) and rand.Next(2, 2) independently for running to show
the probability of victory of the first and second set, respectively.
rand.Next(1, 3) is the randomization between these two sets.